Inherit process namespaces from ancestors
During testing sometimes it's not possible to get a handle on a process before
some aspect of the test has run. This is true of LiveViews, which are spawned
with `{:ok, view, _html} = live(conn, ~p"/some/path")` and are already rendered
by the time we have a handle on the `view` process. Any Snap calls during
`mount/3` or `handle_params/3` have no process namespace set.

To make this kind of thing easier to test, we now use process namespaces to
persist the namespace, and walk up the process tree to find any namespaces
defined on parent processes.

To do this we use ProcessTree, which handles some of the complexity of walking
the tree and the behaviour of different OTP versions.